% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distance.R
\name{compute_distance}
\alias{compute_distance}
\title{Compute Euclidian distance matrix.}
\usage{
compute_distance(x1, x2 = NULL)
}
\arguments{
\item{x1}{Matrix containing coordinates for the first set of locations.}

\item{x2}{Matrix containing coordinates for the second set of locations.}
}
\value{
A matrix with Euclidean distances between two sets of coordinates
  \code{x1} and \code{x2}.  If the second set of coordinates \code{x2} is not
  specified then a matrix pairwise distances for the first set of coordinates
  \code{x1} are returned.
}
\description{
Compute Euclidian distance matrix.
}
\examples{
x1 <- matrix(1:6, ncol =2, byrow = TRUE)
x2 <- matrix(7:12, ncol = 2, byrow = TRUE)
compute_distance(x1)
compute_distance(x1,x2)

}
\author{
Eric W. Fox
}
