% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rmse.R
\name{compute_rmse}
\alias{compute_rmse}
\title{Compute root-mean-square error}
\usage{
compute_rmse(y, y_pred)
}
\arguments{
\item{y}{Numeric data vector.}

\item{y_pred}{Numeric vector of predictions.}
}
\value{
RMSE
}
\description{
Compute root-mean-square error
}
\author{
Eric Fox
}
