% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage}
\alias{coverage}
\title{Check coverage of prediction intervals}
\usage{
coverage(y, pred, se, cl = 0.9)
}
\arguments{
\item{y}{Numeric data vector.}

\item{pred}{Numeric vector of predictions.}

\item{se}{Numeric vector of prediction SEs.}

\item{cl}{Confidence level (default is 0.9).}
}
\value{
Coverage for prediction intervals.
}
\description{
Check coverage of prediction intervals
}
\author{
Eric W. Fox
}
