% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage2.R
\name{coverage2}
\alias{coverage2}
\title{Check coverage of prediction intervals}
\usage{
coverage2(y, ci_lb, ci_ub)
}
\arguments{
\item{y}{numeric data vector}

\item{ci_lb}{numveric vector of lower bounds for prediction intervals}

\item{ci_ub}{numberic vector of upper bounds for prediction intervals}
}
\value{
Coverage for prediction intervals.
}
\description{
In this version user must supply prediction intervals (e.g., from 
\code{lm.predict})
}
\author{
Eric W. Fox
}
