% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rfrk.R
\name{fit_rfrk}
\alias{fit_rfrk}
\title{Fit RFRK}
\usage{
fit_rfrk(rf, y, dist_mtx, theta_ini = NULL)
}
\arguments{
\item{rf}{random forest model}

\item{y}{Numeric response vector.}

\item{dist_mtx}{Distance matrix used to compute covariance matrix.}

\item{theta_ini}{numeric vector. Initial values for covariance parameters:
logarithm transformed nugget, partial sill, and range, in that order.  If
not specified uses default initialization.}
}
\value{
object of type `rfrk'.
}
\description{
Estimate a random forest residual kriging model
}
\note{
assumes exponential covariance function
}
