% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_splm.R
\name{fit_splm}
\alias{fit_splm}
\title{Estimate a spatial linear model}
\usage{
fit_splm(X, y, cov_type, theta_ini = NULL, dist_mtx = NULL,
  dist_K = NULL, dist_S = NULL, est_meth, include_hessian = F,
  knot_coord = NULL)
}
\arguments{
\item{X}{Design matrix.  Must include column of 1's for intercept.}

\item{y}{Numeric response vector.}

\item{cov_type}{character.  Type of covariance model: \code{exponential},
\code{gaussian}, or \code{spherical}.}

\item{theta_ini}{numeric vector. Initial values for covariance parameters:
logarithm transformed nugget, partial sill, and range, in that order.  If
not specified uses default initialization.}

\item{dist_mtx}{Distance matrix used to compute covariance matrix.}

\item{dist_K}{Distance matrix for knot locations.}

\item{dist_S}{Distance matric between data locations and knots (\code{nrow}
is number of data points and \code{ncol} is number of knots).}

\item{est_meth}{character. Estimation method: 'REML' for restricted maximum
likelihood, or 'ML' for maximum likelihood.}

\item{include_hessian}{logical.  Whether or not to compute hessian matrix
when running \code{optim}.}

\item{knot_coord}{data frame or matrix. Coordinates of knots used for reduced
rank method. Optional to include (coordinates are just saved in \code{splm}
object for later reference).}
}
\value{
List of class \code{splm} containg features of fitted spatial linear
  model (e.g., covariance matrix, predictor coefficients, ...).
}
\description{
Estimate parameters for spatial linear model and form object of class
\code{splm}.
}
\details{
The purpose of this function is to automate the process of fitting a spatial
model using \code{optim} and forming an object of class \code{splm}, which
contains the attributes of that fitted model, using \code{make_splm}.
}
\author{
Eric W. Fox
}
