% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_sim.R
\name{geo_sim}
\alias{geo_sim}
\title{Simulate geostatistical data}
\usage{
geo_sim(coord, cov_type = "exponential", nugg, parsil, range)
}
\arguments{
\item{coord}{data.frame with x,y coordinates for simulation.}

\item{cov_type}{character.  Type of covariance model: \code{exponential},
\code{gaussian}, or \code{spherical}; default is \code{exponential}.}

\item{nugg}{nugget parameter}

\item{parsil}{partial sill parameter}

\item{range}{range parameter}
}
\value{
numeric vector with simulated geostatistical data.
}
\description{
Function that simulates zero-mean, spatially autocorrelated data with nugget
effect.  Simulation procedure uses the Cholesky decomposition method
(Cressie, 1993, p. 201).
}
\author{
Eric W. Fox
}
