% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2LL_sk.R
\name{m2LL_sk}
\alias{m2LL_sk}
\title{Minus 2 time the log-likelihood for zero-mean simple kriging}
\usage{
m2LL_sk(theta, r, dist_mtx, boundRange = T)
}
\arguments{
\item{theta}{Numeric vector of covariance parameters (use log transformed
nugget, partial sill, and range, in that order).}

\item{r}{Numeric vector of residuals}

\item{boundRange}{whether to bound effective range at max distance.}
}
\value{
Minus 2 times the log-likelihood for RFRK model.
}
\description{
log-likelihood for residual rfrk model
}
