% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2LL_smw.R
\name{m2LL_smw}
\alias{m2LL_smw}
\title{Minus 2 time the log-likelihood for reduced rank spatial linear model.}
\usage{
m2LL_smw(theta, y, X, dist_K, dist_S, est_meth, cov_type = "exponential")
}
\arguments{
\item{theta}{Numeric vector of covariance parameters (use log transformed
nugget, partial sill, and range, in that order).}

\item{y}{Numeric response vector.}

\item{X}{Design matrix.}

\item{dist_K}{distance matrix for knot locations.}

\item{dist_S}{distance matric between data locations and knots (\code{nrow}
is number of data points and \code{ncol} is number of knots).}

\item{est_meth}{character: Estimation method: 'REML' for restricted maximum
likelihood (default), or 'ML' for maximum likelihood.}

\item{cov_type}{character; type of covariance type of covariance model:
\code{exponential}, \code{gaussian}, or \code{spherical}; default is
\code{exponential}.}
}
\value{
Minus 2 times the log-likelihood for reduced rank spatial linear
  model using Sherman-Morrison-Woodbury formula.  See Cressie FRK paper
  equation 2.16 for decomposition.
}
\description{
Minus 2 time the log-likelihood for reduced rank spatial linear model.
}
\author{
Eric W. Fox
}

