% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_covmat.R
\name{make_covmat}
\alias{make_covmat}
\title{Make covariance matrix.}
\usage{
make_covmat(theta, dist_mtx, is_log, include_nugg, type = "exponential")
}
\arguments{
\item{theta}{numeric vector with 3 covariance parameters: nugget, partial
sill, and range (in that order).}

\item{dist_mtx}{Matrix of pairwise distances between data locations.}

\item{is_log}{logical. If \code{TRUE} then the values of \code{theta} are the
natural logarithms of the parameters.}

\item{include_nugg}{logical. Whether or not to include the nugget along the
diagonal.}

\item{type}{character. Type of covariance model: \code{exponential},
\code{gaussian}, or \code{spherical}; default is \code{exponential}.}
}
\value{
Covariance matrix.
}
\description{
Make covariance matrix with different types of covariance models.
}
\author{
Eric W. Fox
}
