% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_splm.R
\name{make_splm}
\alias{make_splm}
\title{Make a spatial linear model}
\usage{
make_splm(params, X, y, cov_type, dist_mtx = NULL, m2loglike = NULL,
  est_meth = NULL, hessian = NULL, dist_K = NULL, dist_S = NULL,
  knot_coord = NULL, optim_out = NULL)
}
\arguments{
\item{params}{Numeric vector of fitted covariance parameter values (nugget,
partial sill, and range, in that order).}

\item{X}{Design matrix.  Must include column of 1's for intercept.}

\item{y}{Numeric response vector.}

\item{cov_type}{character.  Type of covariance model: \code{exponential},
\code{gaussian}, or \code{spherical}.}

\item{dist_mtx}{Distance matrix used to compute covariance matrix.}

\item{m2loglike}{Minus 2 times the log-likelihood (from \code{optim()}).}

\item{est_meth}{character. Estimation method used for \code{m2LL}: 'REML' for
restricted maximum likelihood, or 'ML' for maximum likelihood.}

\item{hessian}{Hessian matrix from \code{optim()}.}

\item{dist_K}{Distance matrix for knot locations.}

\item{dist_S}{Distance matric between data locations and knots (\code{nrow}
is number of data points and \code{ncol} is number of knots).}

\item{knot_coord}{data frame or matrix. Coordinates of knots used for reduced
rank method.  Optional to include (coordinates are just saved in 'splm'
object for later reference).}

\item{optim_out}{list output from \code{optim()}.}
}
\value{
List of class \code{splm} containg features of fitted spatial linear
  model (e.g., covariance matrix, predictor coefficients, ...).
}
\description{
Make spatial linear model object using estimated covariance parameters from
\code{optim()}.
}
\seealso{
\code{\link{m2LL}}, \code{\link[stats]{optim}}
}
\author{
Eric W. Fox
}
