% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmi_data.R
\docType{data}
\name{mmi_data}
\alias{mmi_data}
\title{Macroinvertebrate Multimetric Index (MMI) Data Set}
\usage{
data(mmi_data)
}
\description{
A data frame containing the MMI scores and condition classes from the NRSA 2008/2009 survey,
  as well as additional site information.  The variables are described below.
}
\details{
\itemize{
  \item NRSA_ID site identification
  \item MMI_BENT. MMI score
  \item BENT_MMI_COND. condition class coded as Good, Fair, and Poor
  \item AGG_ECO9_2015. aggregated 9 ecoregion
  \item STATE
  \item ALB_X. albers x-coord
  \item ALB_Y. albers y-coord
  \item LON_DD83. longitude
  \item LAT_DD83. latitude
  \item Year. year sample collected
  \item DATE_COL. date sample collected
}
}
