% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_krige_batch.R
\name{pred_krige_batch}
\alias{pred_krige_batch}
\title{Kriging predictions and errors}
\usage{
pred_krige_batch(object, obsv_coord, pred_coord, Xp = NULL,
  row_names = NULL, computeSE = TRUE, scale = 1)
}
\arguments{
\item{object}{object of class 'splm'}

\item{obsv_coord}{Data frame  or matrix with Alber's x,y coordinates for
observations.}

\item{pred_coord}{Data frame or matrix with Alber's x,y coordinates for
prediction.}

\item{Xp}{matrix of covariates at prediction sites for universal kriging.
Make sure that the first column of this matrix contains 1's for the
intercept. The default is \code{NULL}, which assumes a spatial model with
no predictors (the function will generate a matrix with a column of 1's for
the intercept).}

\item{row_names}{character or numeric vector specify row names for the
predictions}

\item{computeSE}{logical; whether to compute standard errors, default is
\code{TRUE}}

\item{scale}{scaling factor for distance matrix.  The distance matrix will be
divided by this number. Dafault is 1 for no scaling.  Useful when
converting Alber's coordinates from meters to kilometers; in this case, set
scale to 1000.}
}
\value{
Matrix with kriging predictions and standard errors.
}
\description{
Function that makes krigging predictions using a fitted spatial linear model
and an external prediction data set.
}
\author{
Jay Ver Hoef (edited by Eric Fox)
}
