% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_rfrk.R
\name{pred_rfrk}
\alias{pred_rfrk}
\title{Kriging predictions and SEs for RFRK model}
\usage{
pred_rfrk(rf, rfrk, obsv_coord, pred_coord, Xp, row_names = NULL,
  computeSE = T, scale = 1)
}
\arguments{
\item{rf}{object of class 'randomForest'}

\item{rfrk}{object of class 'rfrk'}

\item{obsv_coord}{Data frame  or matrix with x,y coordinates for
observations.}

\item{pred_coord}{Data frame or matrix with x,y coordinates for predictions.}

\item{Xp}{matrix of covariates at prediction sites.}

\item{row_names}{character or numeric vector specify row names for the
predictions}

\item{computeSE}{logical; whether to compute standard errors, default is
\code{TRUE}}

\item{scale}{scaling factor for distance matrix.  The distance matrix will be
divided by this number. Dafault is 1 for no scaling.  Useful when
converting Alber's coordinates from meters to kilometers; in this case, set
scale to 1000.}
}
\value{
Matrix with RFRK predictions and SEs.
}
\description{
Evaluate kriging predictions and standard errors for random forest regression
kriging model
}
