% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.splm.R
\name{print.splm}
\alias{print.splm}
\title{Print method of spatial linear model}
\usage{
\method{print}{splm}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \code{splm}}

\item{digits}{how many significant digits to be used for numeric values. 
Note that at least 2 places after the decimal will always be printed.}
}
\description{
Print method of spatial linear model
}
\author{
Eric W. Fox
}
