% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.splm.R
\name{print.summary.splm}
\alias{print.summary.splm}
\title{Print method of spatial linear model}
\usage{
\method{print}{summary.splm}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary.splm}}
}
\description{
Print method of spatial linear model
}
\author{
Eric W. Fox
}
