% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamcat_preds.R
\docType{data}
\name{streamcat_preds}
\alias{streamcat_preds}
\title{StreamCat Predictors}
\format{A data frame with StreamCat predictor variables and the NRSA site
  identification \code{NRSA_ID}.}
\source{
\url{http://www.epa.gov/national-aquatic-resource-surveys/streamcat}
}
\usage{
data(streamcat_preds)
}
\description{
StreamCat predictor data for 2008/2009 NRSA sites.
}
