% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.splm.R
\name{summary.splm}
\alias{summary.splm}
\title{Summarize Spatial Linear}
\usage{
\method{summary}{splm}(object, ...)
}
\arguments{
\item{object}{an object of class \code{splm}}
}
\value{
data frame with spatial regression summary of class 'summary.splm'
}
\description{
\code{summary} method for class 'splm'
}
